Split Player Code into Separate Components, Reorganized Scripts Folder
