+++ Assets/Scenes/attack test scene.unity
@@ -146,7 +146,7 @@ GameObject:
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 1
+  m_IsActive: 0
 --- !u!70 &49909081
 CapsuleCollider2D:
   m_ObjectHideFlags: 0
@@ -243,14 +243,14 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 49909080}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: -1.7059474, y: 0.14390534, z: 0}
+  m_LocalPosition: {x: -13.96, y: 2.48, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children:
   - {fileID: 1767731338}
   - {fileID: 890629188}
   - {fileID: 2038987004}
   m_Father: {fileID: 0}
-  m_RootOrder: 1
+  m_RootOrder: 2
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!114 &49909085
 MonoBehaviour:
@@ -264,7 +264,6 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: cedc94de4572a5e46ba3ed2b950553ce, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  attackInput: 106
   attackCD: 0
 --- !u!95 &49909086
 Animator:
@@ -299,6 +298,7 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   playerRB: {fileID: 0}
   spriteR: {fileID: 0}
+  playerController: {fileID: 0}
   isLeft: 1
   playerMovementSpeed: 5
   isGrounded: 0
@@ -324,6 +324,7 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   playerAttack: {fileID: 0}
   playerMovement: {fileID: 0}
+  playerController: {fileID: 0}
   _isMove: 1
 --- !u!114 &49909089
 MonoBehaviour:
@@ -337,9 +338,12 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: d60c9db5f7fcb144eb8b29fa2d10be84, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  input0: 106
-  input1: 107
-  input2: 32
+  playerInput: {fileID: 11400000, guid: 4e35b7bdb3771434fa74007b6ed719b0, type: 2}
+  inputAxis: {x: 0, y: 0}
+  inputJumpDown: 0
+  inputJump: 0
+  input0Down: 0
+  input1Down: 0
 --- !u!1 &65668888
 GameObject:
   m_ObjectHideFlags: 0
@@ -483,7 +487,7 @@ Transform:
   m_LocalScale: {x: 2.246, y: 1.7348, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
-  m_RootOrder: 4
+  m_RootOrder: 5
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &438606347
 GameObject:
@@ -592,7 +596,7 @@ Transform:
   m_LocalScale: {x: 19.963074, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
-  m_RootOrder: 2
+  m_RootOrder: 3
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &519420028
 GameObject:
@@ -797,7 +801,7 @@ Transform:
   m_LocalScale: {x: 1.9471, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
-  m_RootOrder: 3
+  m_RootOrder: 4
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!114 &533292078
 MonoBehaviour:
@@ -1068,7 +1072,7 @@ Transform:
   m_LocalScale: {x: 10.281076, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
-  m_RootOrder: 5
+  m_RootOrder: 6
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!114 &949424311
 MonoBehaviour:
@@ -1083,6 +1087,75 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   hp: 1
+--- !u!1001 &1493049561
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: 3829972687822279505, guid: 8c5eaeb3c8b9f1e4fba821edf9847f2f, type: 3}
+      propertyPath: m_Color.b
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 3829972687822279505, guid: 8c5eaeb3c8b9f1e4fba821edf9847f2f, type: 3}
+      propertyPath: m_Color.g
+      value: 0.8992183
+      objectReference: {fileID: 0}
+    - target: {fileID: 3829972687822279505, guid: 8c5eaeb3c8b9f1e4fba821edf9847f2f, type: 3}
+      propertyPath: m_Color.r
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3829972687822279506, guid: 8c5eaeb3c8b9f1e4fba821edf9847f2f, type: 3}
+      propertyPath: m_Name
+      value: PlayerSquare
+      objectReference: {fileID: 0}
+    - target: {fileID: 3829972687822279510, guid: 8c5eaeb3c8b9f1e4fba821edf9847f2f, type: 3}
+      propertyPath: m_RootOrder
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 3829972687822279510, guid: 8c5eaeb3c8b9f1e4fba821edf9847f2f, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 0.5538609
+      objectReference: {fileID: 0}
+    - target: {fileID: 3829972687822279510, guid: 8c5eaeb3c8b9f1e4fba821edf9847f2f, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: -0.15251791
+      objectReference: {fileID: 0}
+    - target: {fileID: 3829972687822279510, guid: 8c5eaeb3c8b9f1e4fba821edf9847f2f, type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3829972687822279510, guid: 8c5eaeb3c8b9f1e4fba821edf9847f2f, type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 3829972687822279510, guid: 8c5eaeb3c8b9f1e4fba821edf9847f2f, type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3829972687822279510, guid: 8c5eaeb3c8b9f1e4fba821edf9847f2f, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3829972687822279510, guid: 8c5eaeb3c8b9f1e4fba821edf9847f2f, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3829972687822279510, guid: 8c5eaeb3c8b9f1e4fba821edf9847f2f, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3829972687822279510, guid: 8c5eaeb3c8b9f1e4fba821edf9847f2f, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3829972687822279510, guid: 8c5eaeb3c8b9f1e4fba821edf9847f2f, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: 8c5eaeb3c8b9f1e4fba821edf9847f2f, type: 3}
 --- !u!1 &1767731337
 GameObject:
   m_ObjectHideFlags: 0
