+++ Assets/Scripts/MonoBehaviors.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3874200d5ac05c541b1a208d06d7abff
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
