+++ Assets/Scripts/MonoBehaviors/AttackHorizontal.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 67ff86e48a57771488ee2b39be9f27a8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
