+++ Assets/Scripts/MonoBehaviors/Enemy.cs
@@ -1,0 +1,7 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class Enemy : Destructible
+{
+}
