+++ Assets/Scripts/PlayerInputSO.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ac90b96360d67054fad3ec53e5b2fc5c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
