+++ Assets/Scripts/ScriptableObjects/PlayerInput.asset
@@ -1,0 +1,19 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: ac90b96360d67054fad3ec53e5b2fc5c, type: 3}
+  m_Name: PlayerInput
+  m_EditorClassIdentifier: 
+  key0: 106
+  key1: 32
+  keyJump: 107
+  axisH: Horizontal
+  axisV: Vertical
