+++ Assets/Scripts/ScriptableObjects/PlayerInput.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4e35b7bdb3771434fa74007b6ed719b0
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
