+++ UserSettings/EditorUserSettings.asset
@@ -6,10 +6,10 @@ EditorUserSettings:
   serializedVersion: 4
   m_ConfigSettings:
     RecentlyUsedScenePath-0:
-      value: 22424703114646680e0b0227036c521f001e162d6c3c232339690e30e7ee3176f7e93ffdfe
+      value: 22424703114646680e0b0227036c6c111b07142f1f2b233e2867083debf42d
       flags: 0
     RecentlyUsedScenePath-1:
-      value: 22424703114646680e0b0227036c6c111b07142f1f2b233e2867083debf42d
+      value: 22424703114646680e0b0227036c521f001e162d6c3c232339690e30e7ee3176f7e93ffdfe
       flags: 0
     RecentlyUsedScenePath-2:
       value: 22424703114646680e0b0227036c5e0402161b216c3c232339690e30e7ee3176f7e93ffdfe
